// lib/types.ts
export interface Clip {
  id: string;
  title: string;
  titleEn: string;
  titleId: string;
  cover: string;
  thumbnail: string;
  bvid: string;
  startTime: number;
  duration: string;
  tags: string[];
  focus: string;
  focusEn: string;
  focusId: string;
  source: string;
  uploader: string;
}